backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.16
psycopg2-binary==2.9.6
pydantic==1.10.7
pyjwt==2.6.0
bcrypt==4.0.1
python-dotenv==0.21.1
aiohttp==3.8.4
requests==2.28.2
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.10.0
httpx==0.24.1
docker==6.0.1
